refactor(charts): use `index.tsx` as component root file
