+++ packages/charts/src/components/BarChart/BarChart.cy.tsx
@@ -1,5 +1,5 @@
 import { complexDataSet } from '../../resources/DemoProps.js';
-import { BarChart } from './BarChart.js';
+import { BarChart } from './index.js';
 import { cypressPassThroughTestsFactory, testChartLegendConfig, testChartZoomingTool } from '@/cypress/support/utils';
 
 const dimensions = [
