+++ packages/charts/src/components/BarChart/index.tsx
