+++ packages/charts/src/components/ColumnChartWithTrend/index.tsx
