+++ packages/charts/src/components/DonutChart/DonutChart.cy.tsx
@@ -1,5 +1,5 @@
 import { complexDataSet, simpleDataSet } from '../../resources/DemoProps.js';
-import { DonutChart } from './DonutChart.js';
+import { DonutChart } from './index.js';
 import { cypressPassThroughTestsFactory, testChartLegendConfig } from '@/cypress/support/utils';
 
 const dimension = {
