+++ packages/charts/src/components/DonutChart/DonutChart.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { useEffect, useState } from 'react';
 import { legendConfig, simpleDataSet, simpleDataSetWithSmallValues, tooltipConfig } from '../../resources/DemoProps.js';
-import { DonutChart } from './DonutChart.js';
+import { DonutChart } from './index.js';
 
 const meta = {
   title: 'DonutChart',
