+++ packages/charts/src/components/DonutChart/index.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { forwardRef } from 'react';
-import type { PieChartProps } from '../PieChart/PieChart.js';
-import { PieChart } from '../PieChart/PieChart.js';
+import type { PieChartProps } from '../PieChart/index.js';
+import { PieChart } from '../PieChart/index.js';
 
 /**
  * A `DonutChart` is a type of graph that displays data in a circular graph with a cut-out hole in the middle.
