+++ packages/charts/src/components/LineChart/index.tsx
