+++ packages/charts/src/components/PieChart/PieChart.cy.tsx
@@ -1,6 +1,6 @@
 import { Text as RechartsText } from 'recharts';
 import { complexDataSet, simpleDataSet } from '../../resources/DemoProps.js';
-import { PieChart } from './PieChart.js';
+import { PieChart } from './index.js';
 import { cypressPassThroughTestsFactory, testChartLegendConfig } from '@/cypress/support/utils';
 
 const dimension = {
