+++ packages/charts/src/components/RadarChart/RadarChart.cy.tsx
@@ -1,5 +1,5 @@
 import { complexDataSet } from '../../resources/DemoProps.js';
-import { RadarChart } from './RadarChart.js';
+import { RadarChart } from './index.js';
 import { cypressPassThroughTestsFactory, testChartLegendConfig } from '@/cypress/support/utils';
 
 const dimensions = [
