+++ packages/charts/src/components/RadialChart/RadialChart.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { FlexBox, FlexBoxAlignItems, FlexBoxJustifyContent, Text } from '@ui5/webcomponents-react';
 import { ThemingParameters } from '@ui5/webcomponents-react-base';
-import { RadialChart } from './RadialChart.js';
+import { RadialChart } from './index.js';
 
 const meta = {
   title: 'RadialChart',
