+++ packages/charts/src/components/ScatterChart/ScatterChart.cy.tsx
@@ -1,5 +1,5 @@
 import { complexDataSet, scatterComplexDataSet } from '../../resources/DemoProps.js';
-import { ScatterChart } from './ScatterChart.js';
+import { ScatterChart } from './index.js';
 import { cypressPassThroughTestsFactory, testChartLegendConfig } from '@/cypress/support/utils';
 
 const measures = [
