+++ packages/charts/src/index.ts
@@ -1,24 +1,24 @@
-import { BarChart } from './components/BarChart/BarChart.js';
+import { BarChart } from './components/BarChart/index.js';
 import { BarChartPlaceholder } from './components/BarChart/Placeholder.js';
-import { BulletChart } from './components/BulletChart/BulletChart.js';
+import { BulletChart } from './components/BulletChart/index.js';
 import { BulletChartPlaceholder } from './components/BulletChart/Placeholder.js';
-import { ColumnChart } from './components/ColumnChart/ColumnChart.js';
+import { ColumnChart } from './components/ColumnChart/index.js';
 import { ColumnChartPlaceholder } from './components/ColumnChart/Placeholder.js';
-import { ColumnChartWithTrend } from './components/ColumnChartWithTrend/ColumnChartWithTrend.js';
+import { ColumnChartWithTrend } from './components/ColumnChartWithTrend/index.js';
 import { ColumnChartWithTrendPlaceholder } from './components/ColumnChartWithTrend/Placeholder.js';
 import { ComposedChart } from './components/ComposedChart/index.js';
 import { ComposedChartPlaceholder } from './components/ComposedChart/Placeholder.js';
-import { DonutChart } from './components/DonutChart/DonutChart.js';
-import { LineChart } from './components/LineChart/LineChart.js';
+import { DonutChart } from './components/DonutChart/index.js';
+import { LineChart } from './components/LineChart/index.js';
 import { LineChartPlaceholder } from './components/LineChart/Placeholder.js';
-import { PieChart } from './components/PieChart/PieChart.js';
+import { PieChart } from './components/PieChart/index.js';
 import { PieChartPlaceholder } from './components/PieChart/Placeholder.js';
-import { RadarChart } from './components/RadarChart/RadarChart.js';
-import { RadialChart } from './components/RadialChart/RadialChart.js';
+import { RadarChart } from './components/RadarChart/index.js';
+import { RadialChart } from './components/RadialChart/index.js';
+import { ScatterChart } from './components/ScatterChart/index.js';
 import { ScatterChartPlaceholder } from './components/ScatterChart/Placeholder.js';
-import { ScatterChart } from './components/ScatterChart/ScatterChart.js';
+import { TimelineChart } from './components/TimelineChart/index.js';
 import { TimelineChartPlaceholder } from './components/TimelineChart/Placeholder.js';
-import { TimelineChart } from './components/TimelineChart/TimelineChart.js';
 import { TimelineChartAnnotation } from './components/TimelineChart/TimelineChartAnnotation.js';
 import VersionInfo from './generated/VersionInfo.js';
